Refactor chat ui
