+++ instagram/__init__.py
@@ -1,10 +1,9 @@
 from .auth import login, logout
-from .chat_ui import start_chat
-from .commands import CommandRegistry
+from .chat import start_chat
 from .configs import Config
 from .client import cleanup
 
-__all__ = ["login", "logout", "start_chat", "CommandRegistry", "Config", "cleanup"]
+__all__ = ["login", "logout", "start_chat", "Config", "cleanup"]
 
 # import logging
 # logging.basicConfig(level=logging.ERROR)
